pyproject.toml
[project]
name = "project-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio>=6.5.1",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "protobuf==3.20.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
